LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/multicol-001-expected.txt


PASS scroll-snap-align: start
FAIL scroll-snap-align: end assert_equals: expected 120 but got 108
FAIL scroll-snap-align: center assert_equals: expected 180 but got 228
